Move graphs into modules
