+++ dashboard/js/index.js
@@ -7,24 +7,9 @@ import heading from './views/heading.js';
 import wpt from './views/wpt.js';
 import lighthouse from './views/lighthouse.js';
 import cloudinary from './views/cloudinary.js';
-
+import { fetchSummary } from './utils.js';
 
 const dialog = document.querySelector('dialog');
-const dialogTitle = document.getElementById('dialog-title');
-const performanceGraph = document.getElementById('performance-graph');
-const loadGraph = document.getElementById('load-graph');
-const bytesInGraph = document.getElementById('bytes-in-graph');
-const requestsGraph = document.getElementById('requests-graph');
-
-async function fetchSummary(fileName) {
-    try {
-        const response = await fetch(`data/summary/${fileName}`);
-        const data = await response.json();
-        return data;
-    } catch(e) {
-        throw new Error(`fetchSummary() ${e}`);
-    }
-}
 
 function bindEvents() {
     document.querySelector('.dashboard').addEventListener('click', handleDashboardClick);
@@ -50,205 +35,11 @@ function handleCloseDialog() {
     });
 }
 
-function renderWebPageTestGraphs(dataset) {
-    dialogTitle.innerHTML = `<h3>${dataset.url}</h3>`;
-
-    window.d3.json(dataset.src, function(data) {
-        let loadData = [[], []];
-        let requestData = [];
-        let bytesInData = [];
-        let d = {
-            height: 200,
-            left: 120,
-            right: 120,
-            width: 700
-        };
-
-        data = window.MG.convert.date(data, 'date', '%Y-%m-%dT%H:%M');
-
-        data.forEach(entry => {
-            const date = entry.date;
-
-            // if there was an error fetching the test, skip the iteration.
-            if (entry.runError && entry.runError.statusCode) {
-                return;
-            }
-
-            // if for some reason WPT returned no data, skip the iteration.
-            if (Object.keys(entry.metrics).length === 0) {
-                return;
-            }
-
-            loadData[0].push({
-                'date': date,
-                'id': entry.id,
-                'value': parseInt(entry.metrics.documentComplete)
-            });
-
-            loadData[1].push({
-                'date': date,
-                'id': entry.id,
-                'value': parseInt(entry.metrics.fullyLoaded)
-            });
-
-            bytesInData.push({
-                'date': date,
-                'id': entry.id,
-                'value': parseInt(entry.metrics.bytesIn)
-            });
-
-            requestData.push({
-                'date': date,
-                'id': entry.id,
-                'value': parseInt(entry.metrics.requests)
-            });
-        });
-
-        window.MG.data_graphic({
-            area: false,
-            title: 'Load times',
-            data: loadData,
-            width: d.width,
-            height: d.height,
-            left: d.left,
-            right: d.right,
-            target: loadGraph,
-            y_label: 'MS',
-            legend: ['Doc Complete', 'Fully Loaded']
-        });
-
-        window.MG.data_graphic({
-            area: false,
-            title: 'Bytes In',
-            data: bytesInData,
-            width: d.width,
-            height: d.height,
-            left: d.left,
-            right: d.right,
-            target: bytesInGraph
-        });
-
-        window.MG.data_graphic({
-            area: false,
-            title: 'Requests',
-            data: requestData,
-            width: d.width,
-            height: d.height,
-            left: d.left,
-            right: d.right,
-            y_label: 'Total',
-            target: requestsGraph
-        });
-
-        window.d3.selectAll('#load-graph path').on('click', function(data) {
-            const id = data.id;
-
-            if (id) {
-                window.location.href = `https://www.webpagetest.org/results.php?test=${id}`;
-            }
-        });
-
-        window.d3.selectAll('#bytes-in-graph rect').on('click', function(data) {
-            const id = data.id;
-
-            if (id) {
-                window.location.href = `https://www.webpagetest.org/results.php?test=${id}`;
-            }
-        });
-
-        window.d3.selectAll('#requests-graph rect').on('click', function(data) {
-            const id = data.id;
-
-            if (id) {
-                window.location.href = `https://www.webpagetest.org/results.php?test=${id}`;
-            }
-        });
-    });
-}
-
-function renderLighthouseGraphs(dataset) {
-    dialogTitle.innerHTML = `<h3>${dataset.url}</h3>`;
-
-    window.d3.json(dataset.src, function(data) {
-        let perfData = [[], [], [], [], []];
-        let d = {
-            height: 200,
-            left: 120,
-            right: 120,
-            width: 700
-        };
-
-        data = window.MG.convert.date(data, 'date', '%Y-%m-%dT%H:%M');
-
-        data.forEach(entry => {
-            const date = entry.date;
-
-            // if there was an error fetching the test, skip the iteration.
-            if (entry.runError && entry.runError.statusCode) {
-                return;
-            }
-
-            // if for some reason WPT returned no data, skip the iteration.
-            if (Object.keys(entry.metrics).length === 0) {
-                return;
-            }
-
-            perfData[0].push({
-                'date': date,
-                'id': entry.id,
-                'value': parseInt(entry.scores.performance)
-            });
-            perfData[1].push({
-                'date': date,
-                'id': entry.id,
-                'value': parseInt(entry.scores.pwa)
-            });
-            perfData[2].push({
-                'date': date,
-                'id': entry.id,
-                'value': parseInt(entry.scores.accessibility)
-            });
-            perfData[3].push({
-                'date': date,
-                'id': entry.id,
-                'value': parseInt(entry.scores.bestpractices)
-            });
-            perfData[4].push({
-                'date': date,
-                'id': entry.id,
-                'value': parseInt(entry.scores.seo)
-            });
-        });
-
-        window.MG.data_graphic({
-            area: false,
-            title: 'Lighthouse Scores',
-            data: perfData,
-            width: d.width,
-            height: d.height,
-            left: d.left,
-            right: d.right,
-            y_label: '%',
-            target: performanceGraph,
-            legend: ['Performance', 'PWA', 'Accessibility', 'Best Practices', 'SEO']
-        });
-
-        window.d3.selectAll('#performance-graph path').on('click', function(data) {
-            const id = data.id;
-
-            if (id) {
-                window.location.href = `https://www.webpagetest.org/lighthouse.php?test=${id}`;
-            }
-        });
-    });
-}
-
 function showDialog(dataset) {
-
     if (dataset.origin === 'wpt') {
-        renderWebPageTestGraphs(dataset);
+        wpt.renderGraph(dataset);
     } else if (dataset.origin === 'lighthouse') {
-        renderLighthouseGraphs(dataset);
+        lighthouse.renderGraph(dataset);
     }
 
     window.dialogPolyfill.registerDialog(dialog);
